Removed test only methods from utils
